Support for oper-only commands.
We keep a list of command structures, each oper command is placed into the
first free structure and then a /userhost is issued.

When the reply from userhost is received and the user verified as an oper
then the command can be executed, otherwise it should be ignored.  Then the
command structure may be cleared and any dynamic memory freed.

We may receive userhost replies with no data, indicating that the nick left
the network before we could check them.  Since we can't tell who it is,
we need to periodically (in this case, every 2 minutes) go through our
command list and remove any commands that have not been executed in that
period.

dissect_time() also moved to misc.c, now that it is used by more than just
the stats functions.

Current limitation: oper commands may have only one parameter.